+++ CHANGELOG.md
@@ -2,6 +2,7 @@ yyyy/mm/dd Version x.y.z
 ------------------------
 - Updated README.md
 - Expose `pause` argument in client and CLI
+- Rename `client.poke` and `Poked` to `client.read_lcd_memory` and `LcdMemory` respectively
 
 
 2025/01/04 Version 1.0.0